pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aiopagerduty"
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["pagerduty", "async"]
authors = [{ name = "Ravi Terala", email = "terala.work@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "asyncio",
  "aiohttp",
  "async_lru",
  "aiohttp-client-cache",
  "pydantic[email]",
  "pyaml",
  "python-dotenv",
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/terala/aiopagerduty#readme"
Issues = "https://github.com/terala/aiopagerduty/issues"
Source = "https://github.com/terala/aiopagerduty"

[tool.hatch.version]
path = "aiopagerduty/__about__.py"


[tool.hatch.envs.default.scripts]


[tool.hatch.envs.test]
dependencies = [
  "coverage[toml]",
  "pytest",
  "pytest-asyncio",
  "pytest-cov",
  "pytest-mock",
  "assertpy",
]
detached = true
dev-mode = true

[tool.hatch.envs.test.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov-report=html --cov=aiopagerduty --cov=tests"
no-cov = "cov --no-cov"

[[tool.hatch.envs.test.matrix]]
python = ["310"]

[tool.hatch.envs.lint]
dependencies = [
  "aiohttp",

  "pylint",
  "pylint-pydantic",
  "mypy",
  "mypy-extensions",
  "types-pyyaml",
  "types-requests",
  "types-urllib3",
  "typing_extensions",
]
detached = true
dev-mode = false

[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive aiopagerduty/*.py"
lint = "pylint --enable-all-extensions aiopagerduty/*.py"
all = ["typing", "lint"]

[tool.hatch.build]
ignore-vcs = true
include = ["aiopagerduty/*.py"]
reproducible = true

[tool.coverage.run]
branch = true
parallel = true
omit = ["aiopagerduty/__about__.py"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
